Ajout nouvelles StockageException dans la partie service (ModifierPizzaService et SupprimerPizzaService)

+++ src/main/java/fr/pizzeria/menu/FermetureService.java
@@ -4,6 +4,7 @@ import java.sql.SQLException;
 import java.util.Scanner;
 
 import fr.pizzeria.dao.IPizzaDao;
+import fr.pizzeria.dao.PizzaJpaDao;
 import fr.pizzeria.exception.StockageException;
 
 public class FermetureService extends MenuService {
@@ -11,7 +12,15 @@ public class FermetureService extends MenuService {
 	@Override
 	public void executeUC(Scanner scanner, IPizzaDao dao) throws StockageException, SQLException {
 		
-		dao.destroyEmFactory();
+		if(dao instanceof PizzaJpaDao){
+			
+			dao.destroyEmFactory();
+
+		} else {
+			
+			System.out.println("PizzaJpaDao n'est pas activé, vous ne pouvez donc pas clôturer l'EntityManagerFactory");
+			
+		}
 
 	}
